refactor(internal-review-gate): value_objects 分割・DRY 違反修正・docstring 整合
